Feat: view, presenter 중간작업 #63
